fixed the test case in dashboardDao
